Create Mx distance same characters
The program must accept a string S as the input. The program must print the character
along with its maximum distance to reach the same character as the output. If two or more
characters have the same maximum distance then the program must print the first occurring
character and its distance as the output. If there are no such characters in the string "S"
then the program must print "-1" as the output

BOUNDARY CONDTIONS:
3<=Length of the S <=10000

INPUT FORMAT:
The first line contains "S"

OUTPUT FORMAT:
The first line contains the character along with the maximum distance.

EXAMPLE INPUT/OUTPUT 1:
INPUT:  
hellomadam

OUTPUT:
m4

EXPLANATIONS:
In the string hellomadam, from the character "l", the distance to reach the same character "l" is "one"
In the string hello Madam, from the character "m",the distance to reach the same character "m" is "four"
In the string hellomadam, from the character "a", the distance to reach the same character "a" is "two"
Hence, the output is m4



TEST CASE 1:
INPUT:
hellomadam

OUTPUT:
m4


TEST CASE 2:
INPUT:
shy

OUTPUT:
-1